+++ components/ProfilePicture.js
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import { Image, Button } from "react-native-elements";
 import * as ImagePicker from "expo-image-picker";
 import Icon from "react-native-vector-icons/FontAwesome";
-import useImagePicker from "../hooks/useImagePicker";
+import useImagePicker from "../hooks/BlogHooks/useImagePicker";
 
 const ProfilePicture = ({ profilePicture, setProfilePicture }) => {
   const { image, pickImage } = useImagePicker();
