+++ screens/App/BlogScreen.js
@@ -3,7 +3,7 @@ import { FlatList, View } from "react-native";
 import { StyleSheet } from "react-native";
 import { auth } from "../../utils/Firebase";
 import Post from "../../components/Post";
-import useFetchPosts from "../../hooks/useFetchPosts";
+import useFetchPosts from "../../hooks/BlogHooks/useFetchPosts";
 export default function Blog() {
   const { posts, refreshing, fetchPosts } = useFetchPosts(auth.currentUser.uid, true);
